asserting data has changed
