+++ src/js/pages/Review.page.js
@@ -1,5 +1,6 @@
 'use strict'
 const Page = require('./page')
+const { validateTableByCaption } = require('../utils/table-utils')
 
 class ReviewPage extends Page {
   get url () {
@@ -11,22 +12,12 @@ class ReviewPage extends Page {
     await clickCancel.click()
   }
 
-  async validateTableByCaption (captionText, dataTable) {
-    // Find the table by its caption
-    const table = await $(`caption*=${captionText}`).parentElement()
+  async validateActivitiesTable (dataTable) {
+    validateTableByCaption('Rivers fished', dataTable)
+  }
 
-    // Assert the row count
-    const expectedRowCount = dataTable.raw().length - 1
-    const rows = await table.$$('tbody tr')
-    expect(rows.length).toEqual(expectedRowCount)
-
-    // Assert the data in the table
-    const expectedRows = dataTable.raw().slice(1) // Skip the first row (headers), we just want to validate the content
-    const tbodyText = await table.$('tbody').getText() // Convert the whole table to text
-    for (const row of expectedRows) {
-      const rowText = row.filter(cell => cell !== '<any>').join(' ')
-      expect(tbodyText).toContain(rowText) // Assert text in table contains the specified row, because the order of the table can be random
-    }
+  async validateLargeCatchesTable (dataTable) {
+    validateTableByCaption('Salmon and large adult sea trout', dataTable)
   }
 }
