+++ src/js/step_definitions/4-Summary/I_expect_the_summary_page_to_show_the_following_small_catches_###.js
@@ -3,5 +3,5 @@ const { defineStep } = require('@cucumber/cucumber')
 const Summary = require('../../pages/Summary.page')
 
 defineStep('I expect the summary page to show the following small catches', async function (smallCatchesTable) {
-  await Summary.validateTableByCaption('Small adult sea trout (1lb and under)', smallCatchesTable, [6, 7])
+  await Summary.validateSmallCatchTable(smallCatchesTable)
 })
