+++ src/js/step_definitions/9-Review/I_expect_the_review_page_to_show_###.js
@@ -3,9 +3,9 @@ const { defineStep } = require('@cucumber/cucumber')
 const Review = require('../../pages/Review.page')
 
 defineStep('I expect the review page to show the following activities', async function (activityTable) {
-  await Review.validateTableByCaption('Rivers fished', activityTable)
+  await Review.validateActivitiesTable(activityTable)
 })
 
 defineStep('I expect the review page to show the following large catches', async function (activityTable) {
-  await Review.validateTableByCaption('Salmon and large adult sea trout', activityTable)
+  await Review.validateLargeCatchesTable('Salmon and large adult sea trout', activityTable)
 })
